requirements.txt
beautifulsoup4==4.7.1
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.8
django-appconf==1.0.3
django-bootstrap==0.2.4
django-bootstrap4==0.0.8
django-classy-tags==0.6.2
django-cors-headers==3.0.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-youtube-thumbnail==1.0.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pafy==0.3.82
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
requests==2.21.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.1
uritemplate==3.0.0
urllib3==1.24.3
whitenoise==4.1.2
